requirements.txt
Flask==3.1.0
flask_cors==5.0.1
Flask_MySQLdb==2.0.0
PyJWT==2.10.1
pytz==2025.2
gunicorn